feat: editor implement create player delete player  (#52)
Implementation of Create and Delete player:

Player can be added using the button `New Player` which will generate an entity with:
* Entity component + player component + inspectable component + container component

Or you can create manually a player entity using the `New Enity` and you will be able to add manually all the components you want.

* On special note: the player component has the address variable displayed as a read-only variable as well the location.

Fixed also the display of UI buttons for the components.

---------

Signed-off-by: Bernardo Serrano <72756999+Dkrenon@users.noreply.github.com>